Add Win Condition
Full Feature List:
A "Holdout" shader, Acts similar to the one of the same name in Blender.
A script pair that changes the render order of any material or shader at runtime or edit mode.
A script pair that can change the object's collision layer on hitting a trigger
The goal script. Super basic right now, only checks for a tagged object on trigger and prints to the console, but has an event to easily expand functionality
A prefab to easily drag and drop the goal hole into a scene
A set of basic color materials for easy testing and to avoid the blinding white of the default material
A steel material to make the ball look nice
Upped the gravity to -15. Feels better, but can still be changed.