Implement Enhanced Gyro Controls
New GravityController (rename to GyroController?) directly modifies the gravity of the world. A method has been added to reset the gyroscope so that the gravity vector interprets the current attitude as a resting position. 

Since the maze no longer moves, the directional light rotates with the gyroscope. This specific behavior will need modification.