Implemented Tiltable and TiltableCamera
Tiltable and TiltableCamera provide various methods of tilting an object based on the gyroscope. Both components will tilt an object as a factor of the gyroscope and will tilt to a maximum angle accordingly. For example, A Tiltable with a maximum angle of 45 but a tilt factor of 0.5 will require a 90 tilt of the gyroscope to reach its peak. Any axis of rotation can be constrained. A Tiltable can choose to orbit an object instead of rotating about its origin. TiltableCamera works the same as the Tiltable orbit mode, but executes in LateUpdate as a matter of convention. These rotations are applied on top of the initial rotation of the object.

GravityController has been renamed to a more generic "Gyro" and it now exposes the modified tilt of the gyroscope for other classes to use. 
Angles have finally been fixed. Quaternions had a problem using signed angles, but the math performed on the tilt is easier to implement using signed angles. Now when given an Euler Angle, the methods first convert to an signed angle and then return the unsigned angle after modification.

Assorted test objects have been added, including a helix and bumper bowl. The bumper implemented is separate from the one implemented in feature/traps.

General organizational changes. Physics Materials are now a separate folder at the root instead of being a sub-folder of Materials. Scripts pertaining to Gyro controls are now in sub-folder Gyro and scripts that are defined as Modifiers (sort of like Decorators,) are in the sub-folder Modifiers. Editor scripts have been created for the new gyro scripts.