Implement Virtual Gyroscope
A custom editor window that allows you to control the Gyro without a gyroscope. Accessible through Tools > Virtual Gyro. You can drag the mouse in the model view like in the model inspector, this will change pitch and roll. There is a bug with weird rotations when manually adjusting pitch with the slider past 90 degrees in either direction. This shouldn't be an issue for this project though. Currently, there is no way to adjust Yaw in the model view, only with the slider.