Fix Invalid Gyroscope
More often than not, we would attempt to store a value from the gyroscope before it was initialized to a valid quaternion. The initial fix checked for equality to a zeroed value, but that solution didn't work with how Unity checks for equality. Now each component of the quaternion is checked.