Bugfixes
Hopefully squashed the bug with gyroscope input by moving most of the logic into Awake and setting timeScale to zero if a valid gyroscope input wasn't obtained. Start is now only in charge of waiting for that valid quaternion and setting the timeScale back to 1 if the quaternion wasn't valid on Start.

VirtualGyro had an issue with immediately jumping back to the previous position after resetting the orientation. This has been fixed by changing the drag value to zero when reset. The check for lastDir was removed as it turned out to not mean anything.