Replace RenderQueueModifier Function
RenderQueueModifier, although fine in play mode, was a pain to get working in edit mode and overall a bother to work with. This has been replaced custom ShaderGUI that allows the render queue of the standard and standard specular shader to be overridden. RenderQueueModifier and its editor have been removed. Prefabs and materials have been updated, but may still be broken in the scene.