Add Testing Tools
Scripts have been added to make testing easier.
Virtual Gyro Recorder takes the input from VirtualGyro and writes it to a json. This is currently a MonoBehaviour, but it would be ideal if it could be made into an EditorWindow or added to VirtualGyro.
ScriptedGyro takes in a json file made by Virtual Gyro Recorder and injects it into Gyro's VirtualTilt.
Gyro now has a means of forcing use of VirtualTilt, ignoring the input gyroscope if toggled. This only works in the editor and development builds (although we may choose to change this if we want a means of making tutorial levels.)
A Frame Rate Counter has been added. This reports the frame rate over a given refresh rate (default one second) and the average frame rate over the life of the scene. Currently, the average is skewed down due to scene loading.
One instruction set has been included to test with.
Various material and project settings have been modified as part of graphics testing.